Thread pool basically working
The thread pool is basically working. It can run tasks in worker(s)
which amount is determined by current HW (number of CPU cores, HT
support etc.)
There are also mechanisms implemented to notify the worker from the task
that the task processing is waiting for some event. This is going to be
used in dynamic thread pool resize (amount of workers)