Worker State Machine Redesign
Originally the worker had just a state which was quite difficult to
handle in terms of state transitions. So the whole system of worker
state has been redesigned using the State design pattern to handle the
state machine. There are still some modifications to be done but
basically it seems to work.