ThreadPool Worker list redesigned
ThreadPool's worker instances are now stored directly in the workers
list (not unsing std::shared_ptr). This change makes the instances
removal working. The workers which are shut down are now removed using
the removal predicate which eliminates the iteration through the list of
workers (which may lead to SEGFAULT).
Also few other tweaks were done as well. But there are still some more
stuff to fix...